chore: use user and group admin on au scope for replicator sp
Assigning custom role to the replicator on AU scope is not possible due to a bug in Azure

See https://github.com/hashicorp/terraform-provider-azuread/issues/1546.

Therefore we need to assign User and Groups administrator built in roles. This only allows replicator to have admin permissions on the groups and users assigned to the AU.

For replicator to assign users to groups successfully, we therefore need User.Read.All on tenant level.